refactor: purge dead blog files
